for DJ voice status add voice commands

+++ storage/messageSpecial/next/index.js
@@ -12,10 +12,28 @@ async function checkTheTypeOfTheGuild(bot, channel, callback) {
         const type = data.type;
         if (acceptedTypes.includes(type)) callback();
         else {
-            channel.send("C'est fonctionnalité n'est pas disponible pour le momment").then(async (msg) => {
-                await bot.basicFunctions.get("wait").run(10000)
-                if (msg.deletable) msg.delete();
-            })
+            if (["dm", "text"].includes(channel.type)) {
+                channel.send("C'est fonctionnalité n'est pas disponible pour le momment").then(async (msg) => {
+                    await bot.basicFunctions.get("wait").run(10000)
+                    if (msg.deletable) msg.delete();
+                })
+            }
+        }
+    });
+}
+
+async function next(bot, channel, user) {
+    checkTheTypeOfTheGuild(bot, channel, async () => {
+        const meChannelId = channel.guild.me.voice.channelID;
+        const member = await channel.guild.members.fetch(user.id);
+        if (member.voice.channelID === meChannelId) {
+            const voiceState = channel.guild.me.voice;
+            const connection = voiceState.connection;
+            const vcChannel = voiceState.channel;
+            if (connection) {
+                if (connection.dispatcher) connection.dispatcher.destroy();
+                bot.musicFunctions.get("joinVoiceChannel").run(bot, vcChannel.id);
+            }
         }
     });
 }
@@ -23,23 +41,14 @@ async function checkTheTypeOfTheGuild(bot, channel, callback) {
 module.exports.addReaction = async (bot, reaction, user, messageData, index) => {
     reaction.users.remove(user.id);
 
-
-    checkTheTypeOfTheGuild(bot, reaction.message.channel, async () => {
-        const meChannelId = reaction.message.guild.me.voice.channelID;
-        const member = await reaction.message.guild.members.fetch(user.id);
-        if (member.voice.channelID === meChannelId) {
-            const voiceState = reaction.message.guild.me.voice;
-            const connection = voiceState.connection;
-            const channel = voiceState.channel;
-            if (connection) {
-                if (connection.dispatcher) connection.dispatcher.destroy();
-                bot.musicFunctions.get("joinVoiceChannel").run(bot, channel.id);
-            }
-        }
-    });
+    next(bot, reaction.message.channel, user);
 }
 
 module.exports.removeReaction = async (bot, reaction, user, messageData, index) => {}
+
+module.exports.next = async (bot, channel, user) => {
+    next(bot, channel, user);
+}
 
 module.exports.help = {
     name: "index"
